[Refactor] models 분리
