+++ models/__init__.py
@@ -1,1 +1,17 @@
-from .models import User, Record, EmotionCategory  # 필요한 모델 모두 import
+from .user import User
+from .record import Record
+from .category import EmotionCategory, SpendCategory
+from .budget import Budget
+from .report import AdminDailyReport, AdminMonthlyReport, AdminWeeklyReport, UserReport
+
+__all__ = [
+    "User",
+    "Record",
+    "EmotionCategory",
+    "SpendCategory",
+    "Budget",
+    "AdminDailyReport",
+    "AdminMonthlyReport",
+    "AdminWeeklyReport",
+    "UserReport",
+]
